+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added multiple routing keys to a single queue for monitoring.
 
 ### Changed
+- run_program:  Added loop for log information to log all queue name/routing key combinations.
+- run_program:  Changed program lock flavor id to the new configuration settings.
 - config/rabbitmq.py.TEMPLATE:  Changed format to allow for multiple queues.
 - Documentation updates.